Add ALPN-supporting SSLContext wrappers
Use Jetty's library to support ALPN.

The whole thing is somewhat hacky in nature. Jetty's lib has to replace
the Java SSL implementations, and use a custom registry to determine
what SSLEngine-s to interfere with.

Akka's SSL support only takes an SSLContext and creates the SSLEngine
itself, but we need to apply the ALPN construction to the engine.
Unfortunately, mostly SSLContext methods are final, so we need to create
a whole new context implementation that just delegates to a pre-existing
context, but can apply a transformation to the results.